TEIID-2177 adding ldap canned queries

+++ connectors/translator-ldap/src/main/java/org/teiid/translator/ldap/LDAPDirectCreateUpdateDeleteQueryExecution.java
@@ -22,6 +22,7 @@
 package org.teiid.translator.ldap;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.StringTokenizer;
 
@@ -45,27 +46,29 @@ public class LDAPDirectCreateUpdateDeleteQueryExecution implements ProcedureExec
 	protected LDAPExecutionFactory executionFactory;
 	protected ExecutionContext executionContext;
 	private int updateCount = -1;
+	private boolean returnsArray = true;
+	private String query;
 	
-	public LDAPDirectCreateUpdateDeleteQueryExecution(List<Argument> arguments, LDAPExecutionFactory factory, ExecutionContext executionContext, LdapContext connection) {
+	public LDAPDirectCreateUpdateDeleteQueryExecution(List<Argument> arguments, LDAPExecutionFactory factory, ExecutionContext executionContext, LdapContext connection, String query, boolean returnsArray) {
 		this.arguments = arguments;
 		this.executionFactory = factory;
 		this.executionContext = executionContext;
 		this.ldapConnection = connection;
+		this.query = query;
+		this.returnsArray = returnsArray;
 	}
 	
 	@Override
 	public void execute() throws TranslatorException {
-		String query = (String)arguments.get(0).getArgumentValue().getValue();
 		String firstToken = null;
 		
 		StringTokenizer st = new StringTokenizer(query, ";"); //$NON-NLS-1$
 		if (st.hasMoreTokens()) {
 			firstToken = st.nextToken();
-			if (!firstToken.equalsIgnoreCase("create") && !firstToken.equalsIgnoreCase("update") && !firstToken.equalsIgnoreCase("delete")) { //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$
-				throw new TranslatorException(LDAPPlugin.Util.gs(LDAPPlugin.Event.TEIID12009));
-			}
 		}
-		
+		if (firstToken == null || (!firstToken.equalsIgnoreCase("create") && !firstToken.equalsIgnoreCase("update") && !firstToken.equalsIgnoreCase("delete"))) { //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$
+			throw new TranslatorException(LDAPPlugin.Util.gs(LDAPPlugin.Event.TEIID12009));
+		}		
 		LdapContext ldapCtx = null;
 		try {
 			ldapCtx = (LdapContext)this.ldapConnection.lookup("");  //$NON-NLS-1$
@@ -75,6 +78,9 @@ public class LDAPDirectCreateUpdateDeleteQueryExecution implements ProcedureExec
 
 		if (firstToken.equalsIgnoreCase("delete")) { // //$NON-NLS-1$
 			String theDN = getDN(st); // the token after the marker is always DN
+			if (st.hasMoreTokens()) {
+				throw new TranslatorException(LDAPPlugin.Util.gs(LDAPPlugin.Event.TEIID12013, st.nextToken()));
+			}
 			try {
 				ldapCtx.destroySubcontext(theDN);
 				this.updateCount = 1;
@@ -86,7 +92,10 @@ public class LDAPDirectCreateUpdateDeleteQueryExecution implements ProcedureExec
 		}
 		else if (firstToken.equalsIgnoreCase("create")) { //$NON-NLS-1$
 			String theDN = getDN(st); // the token after the marker is always DN
-			ArrayList<BasicAttribute> attributes = getAttributes(st, this.arguments);
+			ArrayList<BasicAttribute> attributes = getAttributes(st);
+			if (st.hasMoreTokens()) {
+				throw new TranslatorException(LDAPPlugin.Util.gs(LDAPPlugin.Event.TEIID12013, st.nextToken()));
+			}
 			BasicAttributes attrs = new BasicAttributes();
 			for (BasicAttribute ba:attributes) {
 				attrs.put(ba);
@@ -102,7 +111,10 @@ public class LDAPDirectCreateUpdateDeleteQueryExecution implements ProcedureExec
 		}
 		else if (firstToken.equalsIgnoreCase("update")) { //$NON-NLS-1$
 			String theDN = getDN(st); // the token after the marker is always DN
-			ArrayList<BasicAttribute> attributes = getAttributes(st, this.arguments);
+			ArrayList<BasicAttribute> attributes = getAttributes(st);
+			if (st.hasMoreTokens()) {
+				throw new TranslatorException(LDAPPlugin.Util.gs(LDAPPlugin.Event.TEIID12013, st.nextToken()));
+			}
 			ModificationItem[] updateMods = new ModificationItem[attributes.size()];
 			int i=0;
 			for (BasicAttribute ba:attributes) {
@@ -117,7 +129,6 @@ public class LDAPDirectCreateUpdateDeleteQueryExecution implements ProcedureExec
 				throw new TranslatorException(e, LDAPPlugin.Util.getString("LDAPUpdateExecution.updateFailedUnexpected",theDN));//$NON-NLS-1$
 			}			
 		}
-		
 	}
 	
 	private String getDN(StringTokenizer st) throws TranslatorException {
@@ -127,26 +138,26 @@ public class LDAPDirectCreateUpdateDeleteQueryExecution implements ProcedureExec
 		return st.nextToken();
 	}
 
-	private ArrayList<BasicAttribute> getAttributes(StringTokenizer st, List<Argument> arguments) throws TranslatorException {
+	private ArrayList<BasicAttribute> getAttributes(StringTokenizer st) throws TranslatorException {
 		if (!st.hasMoreTokens()) {
 			throw new TranslatorException(LDAPPlugin.Util.gs(LDAPPlugin.Event.TEIID12011));
 		}
 		
 		ArrayList<BasicAttribute> attributes = new ArrayList<BasicAttribute>();
 		
-		while(st.hasMoreElements()) {
+		if(st.hasMoreElements()) {
 			String var = st.nextToken();
 			
 			int index = var.indexOf('=');
 			if (index == -1) {
-				continue;
+				throw new TranslatorException(LDAPPlugin.Util.gs(LDAPPlugin.Event.TEIID12011));
 			}
-			String key = var.substring(0, index).trim().toLowerCase();
+			String key = var.substring(0, index).trim();
 			String value = var.substring(index+1).trim();
 			
 			if (key.equalsIgnoreCase(ATTRIBUTES)) {
 				StringTokenizer attrTokens = new StringTokenizer(value, ","); //$NON-NLS-1$
-				int attrCount = 1;
+				int attrCount = 0;
 				while(attrTokens.hasMoreElements()) {
 					String name = attrTokens.nextToken().trim();
 					if (arguments.size() <= attrCount) {
@@ -156,7 +167,8 @@ public class LDAPDirectCreateUpdateDeleteQueryExecution implements ProcedureExec
 					Object  anObj = argument.getArgumentValue().getValue();
 					attributes.add(new BasicAttribute(name, anObj));
 				}
-				
+			} else {
+				throw new TranslatorException(LDAPPlugin.Util.gs(LDAPPlugin.Event.TEIID12011));
 			}
 		}
 		return attributes;
@@ -170,10 +182,12 @@ public class LDAPDirectCreateUpdateDeleteQueryExecution implements ProcedureExec
 	@Override
 	public List<?> next() throws TranslatorException, DataNotAvailableException {
 		if (this.updateCount != -1) {
-			Object[] columns = new Object[1];
-			columns[0] = this.updateCount;
-			List<Object[]> row = new ArrayList<Object[]>(1);
-			row.add(columns);
+			List<Object> row = Arrays.asList((Object)Integer.valueOf(this.updateCount));
+			if (returnsArray) {
+				Object[] columns = new Object[1];
+				columns[0] = this.updateCount;
+				row.set(0, columns);
+			}
 			this.updateCount = -1;
 			return row;
 		}
